fix: 🐛 remove theme from main menu bar and move to settings pag
