+++ addons/core/translations/en-us.yaml
@@ -48,6 +48,7 @@ titles:
   cli: CLI
   cache-daemon: Cache Daemon
   client: Client
+  color-theme: Color Theme
 descriptions:
   empty-set: There are no items to display yet.  You may be able to add items or try back later.
   cluster-url-initialization: To get started, please enter your cluster URL
