+++ ui/desktop/app/routes/application.js
@@ -7,7 +7,6 @@ import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
 import { action } from '@ember/object';
 import { A } from '@ember/array';
-import { getOwner } from '@ember/application';
 import { notifyError } from 'core/decorators/notify';
 
 export default class ApplicationRoute extends Route {
@@ -34,9 +33,13 @@ export default class ApplicationRoute extends Route {
   async beforeModel() {
     this.intl.setLocale(['en-us']);
     await this.session.setup();
-    const theme = this.session.get('data.theme');
-    this.toggleTheme(theme);
     await this.clusterUrl.updateClusterUrl();
+    const theme = this.session.get('data.theme');
+    /* eslint-disable-next-line ember/no-controller-access-in-routes */
+    const controller = this.controllerFor(
+      'scopes.scope.projects.settings.index',
+    );
+    controller.toggleTheme(theme);
 
     // Add token to cache daemon after a successful authentication restoration
     if (this.session.isAuthenticated) {
@@ -94,32 +97,6 @@ export default class ApplicationRoute extends Route {
       return false;
     }
     return true;
-  }
-
-  /**
-   * Applies the specified color theme to the root ember element.
-   * @param {string} theme - "light", "dark", or nullish (system default)
-   */
-  @action
-  toggleTheme(theme) {
-    const rootElementSelector = getOwner(this).rootElement;
-    const rootEl = getOwner(this)
-      .lookup('service:-document')
-      .querySelector(rootElementSelector);
-    this.session.set('data.theme', theme);
-    switch (theme) {
-      case 'light':
-        rootEl.classList.add('rose-theme-light');
-        rootEl.classList.remove('rose-theme-dark');
-        break;
-      case 'dark':
-        rootEl.classList.add('rose-theme-dark');
-        rootEl.classList.remove('rose-theme-light');
-        break;
-      default:
-        rootEl.classList.remove('rose-theme-dark');
-        rootEl.classList.remove('rose-theme-light');
-    }
   }
 
   @action
